Fix typehint for shift setter

+++ sardine/clock/time.py
@@ -60,7 +60,7 @@ class Time(BaseHandler):
         return shift.get()
 
     @shift.setter
-    def shift(self, seconds: int):
+    def shift(self, seconds: float):
         shift.set(seconds)
 
     @contextlib.contextmanager
